Brings Azure object storage more inline with S3 and GCP.
The injected storage service uses a Singleton to construct the
appropriate storage class from the factory. Consequently,
the storage parsing was not using the inherited class in
Azure so the code would fail looking for azure://.
Additionally, the parsing needs to not URL decode the connection string, but the variable to conditionally check
that was not available at the appropriate level.