Brings Azure object storage more inline with S3 and GCP. (#10)
* Brings Azure object storage more inline with S3 and GCP.

The injected storage service uses a Singleton to construct the
appropriate storage class from the factory. Consequently,
the storage parsing was not using the inherited class in
Azure so the code would fail looking for azure://.
Additionally, the parsing needs to not URL decode the connection string, but the variable to conditionally check
that was not available at the appropriate level.

* Removes the optional useUrlDecode parameter; replaces converted spaces to '+' in Azure account key.

* Changes Parameters for protected property to private readonly field to tighten accessibility

* Changes AzureObjectStorage constructor signature to not accept connection string

* Assumes default {name}.blob.core.windows.net unless Endpoint present in connection string

This will simplify the connection string and factory requirements for Azure object storage. Also brings more inline with S3 implementation.

* Remove unnecessary `using`

* Renames private field name; removes protected setter

* Adds container to connection string rather than being extracted during operation calls

---------

Co-authored-by: Robert Emmett <robert.emmett@centeva.com>
Co-authored-by: Pete Deffendol <pete@deffendol.org>