Make storage path strongly typed
Instead of using a string type for storage paths, switch to a non-primitive type StoragePath.  This lets us encapsulate several of the operations needed for paths.

All storage operations now take a StoragePath parameter.  This can be converted to and from a string implicitly, so it's likely that existing code will work without modification, but for safety this should be a major version bump.

+semver:breaking