Allow specifying Content-Disposition header when writing object
When used with a cloud storage provider, it will store the HTTP header
value with the blob such that when downloading via a signed URL or
other mechanism, the header will be sent as part of the response.  This
allows you to specify a filename that is separate from the object name
as stored.

+semver:minor